pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gloaks-cli"
version = "3.0.0"
description = "A comprehensive network reconnaissance and security assessment tool."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Gloaks Team", email = "security@gloaks.io" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiodns>=3.1.1",
    "structlog>=24.1.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "pylint>=3.0.3",
    "mypy>=1.8.0",
]

[project.scripts]
gloaks = "gloaks.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/gloaks"]
